corrected accidental parameter order swap in method

+++ src/main/java/com/qelery/mealmojo/api/model/MenuItem.java
@@ -26,5 +26,5 @@ public class MenuItem {
     private Restaurant restaurant;
 
     @Column(name="restaurant_id", insertable=false, nullable=false, updatable=false)
-    private Integer restaurantId;
+    private Long restaurantId;
 }
